Shifted_Lognormal {brms}	R Documentation

The Shifted Log Normal Distribution

Description

Density, distribution function, quantile function and random generation for the shifted log normal distribution with mean meanlog, standard deviation sdlog, and shift parameter shift.

Usage

dshifted_lnorm(x, meanlog = 0, sdlog = 1, shift = 0, log = FALSE)

pshifted_lnorm(q, meanlog = 0, sdlog = 1, shift = 0, lower.tail = TRUE,
  log.p = FALSE)

qshifted_lnorm(p, meanlog = 0, sdlog = 1, shift = 0, lower.tail = TRUE,
  log.p = FALSE)

rshifted_lnorm(n, meanlog = 0, sdlog = 1, shift = 0)

Arguments

x, q
Vector of quantiles.

meanlog
Vector of means.

sdlog
Vector of standard deviations.

shift
Vector of shifts.

log
Logical; If TRUE, values are returned on the log scale.

lower.tail
Logical; If TRUE (default), return P(X <= x). Else, return P(X > x) .

log.p
Logical; If TRUE, values are returned on the log scale.

p
Vector of probabilities.

n
Number of samples to draw from the distribution.

Details

See vignette("brms_families") for details on the parameterization.

[Package brms version 2.3.0 Index]