StudentT {brms}	R Documentation

The Student-t Distribution

Description

Density, distribution function, quantile function and random generation for the Student-t distribution with location mu, scale sigma, and degrees of freedom df.

Usage

dstudent_t(x, df, mu = 0, sigma = 1, log = FALSE)

pstudent_t(q, df, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qstudent_t(p, df, mu = 0, sigma = 1)

rstudent_t(n, df, mu = 0, sigma = 1)

Arguments

x, q
Vector of quantiles.

df
Vector of degrees of freedom.

mu
Vector of location values.

sigma
Vector of scale values.

log, log.p
Logical; If TRUE, values are returned on the log scale.

lower.tail
Logical; If TRUE (default), return P(X <= x). Else, return P(X > x) .

p
Vector of probabilities.

n
Number of samples to draw from the distribution.

Details

See vignette("brms_families") for details on the parameterization.

See Also

TDist

[Package brms version 2.3.0 Index]