combine_models {brms}	R Documentation

Combine Models fitted with brms

Description

Combine multiple brmsfit objects, which fitted the same model. This is usefuly for instance when having manually run models in parallel.

Usage

combine_models(..., mlist = NULL, check_data = TRUE)

Arguments

...
One or more brmsfit objects.

mlist
Optional list of one or more brmsfit objects.

check_data
Logical; indicates if the data should be checked for being the same across models (defaults to TRUE). Setting it to FALSE may be useful for instance when combining models fitted on multiple imputed data sets.

Details

This function just takes the first model and replaces its stanfit object (slot fit) by the combined stanfit objects of all models.

Value

A brmsfit object.

[Package brms version 2.3.0 Index]