ARR(r) correlation structure

This function is a constructor for the cor_arma class allowing for autoregressive effects of the response only.

cor_arr(formula = ~1, r = 1)

Arguments

formula
A one sided formula of the form ~ t, or ~ t | g, specifying a time covariate t and, optionally, a grouping factor g. A covariate for this correlation structure must be integer valued. When a grouping factor is present in formula, the correlation structure is assumed to apply only to observations within the same grouping level; observations with different grouping levels are assumed to be uncorrelated. Defaults to ~ 1, which corresponds to using the order of the observations in the data as a covariate, and no groups.

r
A non-negative integer specifying the autoregressive response (ARR) order of the ARMA structure. Default is 1.

Value

An object of class cor_arma containing solely autoregressive response terms.

Details

AR refers to autoregressive effects of residuals, which is what is typically understood as autoregressive effects. However, one may also model autoregressive effects of the response variable, which is called ARR in brms.

See also

Examples

cor_arr(~visit|patient, r = 2)

#> arma(~visit|patient, 0, 0, 2)