cor_fixed {brms}	R Documentation

Fixed user-defined covariance matrices

Description

Define a fixed covariance matrix of the response variable for instance to model multivariate effect sizes in meta-analysis.

Usage

cor_fixed(V)

Arguments

V
Known covariance matrix of the response variable. If a vector is passed, it will be used as diagonal entries (variances) and covariances will be set to zero.

Value

An object of class cor_fixed.

Examples

## Not run: 
dat <- data.frame(y = rnorm(3))
V <- cbind(c(0.5, 0.3, 0.2), c(0.3, 1, 0.1), c(0.2, 0.1, 0.2))
fit <- brm(y~1, data = dat, autocor = cor_fixed(V))

## End(Not run)

[Package brms version 2.3.0 Index]