Execute a Function Call

Execute a function call similar to do.call, but without deparsing function arguments.

do_call(what, args, pkg = NULL)

Arguments

what
Either a function or a non-empty character string naming the function to be called.

args
A list of arguments to the function call. The names attribute of args gives the argument names.

pkg
Optional name of the package in which to search for the function if what is a character string.

Value

The result of the (evaluated) function call.