inhaler {brms}	R Documentation

Clarity of inhaler instructions

Description

Ezzet and Whitehead (1991) analyze data from a two-treatment, two-period crossover trial to compare 2 inhalation devices for delivering the drug salbutamol in 286 asthma patients. Patients were asked to rate the clarity of leaflet instructions accompanying each device, using a 4-point ordinal scale.

Usage

inhaler

Format

A dataframe of 572 observations containing information on the following 5 variables.

subject

The subject number

rating

The rating of the inhaler instructions on a scale ranging from 1 to 4

treat

A contrast to indicate which of the two inhaler devices was used

period

A contrast to indicate the time of administration

carry

A contrast to indicate possible carry over effects

Source

Ezzet, F., & Whitehead, J. (1991). A random effects model for ordinal responses from a crossover trial. Statistics in Medicine, 10(6), 901-907.

Examples

## Not run: 
## ordinal regression with family "sratio"
fit1 <- brm(rating ~ treat + period + carry, 
            data = inhaler, family = sratio(), 
            prior = set_prior("normal(0,5)"))
summary(fit1) 
plot(fit1)
       
## ordinal regression with family "cumulative" 
## and random intercept over subjects             
fit2 <- brm(rating ~ treat + period + carry + (1|subject), 
              data = inhaler, family = cumulative(), 
              prior = set_prior("normal(0,5)"))
summary(fit2) 
plot(fit2)

## End(Not run)

[Package brms version 2.3.0 Index]