mo {brms}	R Documentation

Monotonic Predictors in brms Models

Description

Monotonic Predictors in brms Models

Usage

mo(x)

Arguments

x
An integer variable or an ordered factor to be modeled as monotonic.

Details

For detailed documentation see help(brmsformula) as well as vignette("brms_monotonic").

This function is almost solely useful when called in formulas passed to the brms package.

See Also

brmsformula

Examples

  
## Not run: 
# generate some data
income_options <- c("below_20", "20_to_40", "40_to_100", "greater_100")
income <- factor(sample(income_options, 100, TRUE), 
                 levels = income_options, ordered = TRUE)
mean_ls <- c(30, 60, 70, 75)
ls <- mean_ls[income] + rnorm(100, sd = 7)
dat <- data.frame(income, ls)

# fit a simple monotonic model
fit1 <- brm(ls ~ mo(income), data = dat)

# summarise the model
summary(fit1)
plot(fit1, N = 6)
plot(marginal_effects(fit1), points = TRUE)

# model interaction with other variables
dat$x <- sample(c("a", "b", "c"), 100, TRUE)
fit2 <- brm(ls ~ mo(income)*x, data = dat)

# summarise the model
summary(fit2)
plot(marginal_effects(fit2), points = TRUE)

## End(Not run) 
 

[Package brms version 2.3.0 Index]