ngrps.brmsfit {brms}	R Documentation

Number of levels

Description

Extract the number of levels of one or more grouping factors.

Usage

## S3 method for class 'brmsfit'
ngrps(object, ...)

ngrps(object, ...)

Arguments

object
An R object.

...
Currently ignored.

Value

A named list containing the number of levels per grouping factor.

[Package brms version 2.3.0 Index]