nsamples.brmsfit {brms}	R Documentation

Number of Posterior Samples

Description

Extract the number of posterior samples stored in a fitted Bayesian model.

Usage

## S3 method for class 'brmsfit'
nsamples(x, subset = NULL, incl_warmup = FALSE, ...)

nsamples(x, ...)

Arguments

x
An R object

subset
An optional integer vector defining a subset of samples to be considered.

incl_warmup
A flag indicating whether to also count warmup / burn-in samples.

...
Further arguments passed to or from other methods.

Details

Currently there are methods for brmsfit objects.

[Package brms version 2.3.0 Index]