pairs.brmsfit {brms}	R Documentation

Create a matrix of output plots from a brmsfit object

Description

A pairs method that is customized for MCMC output.

Usage

## S3 method for class 'brmsfit'
pairs(x, pars = NA, exact_match = FALSE, ...)

Arguments

x
An object of class brmsfit

pars
Names of parameters for which posterior samples should be returned, as given by a character vector or regular expressions. By default, all posterior samples of all parameters are extracted.

exact_match
Indicates whether parameter names should be matched exactly or treated as regular expression. Default is FALSE.

...
Further arguments to be passed to mcmc_pairs.

Details

For a detailed description see mcmc_pairs.

Examples

## Not run: 
fit <- brm(count ~ log_Age_c + log_Base4_c * Trt_c 
           + (1|patient) + (1|visit), 
           data = epilepsy, family = "poisson")  
pairs(fit, pars = parnames(fit)[1:3], exact_match = TRUE)
pairs(fit, pars = "^sd_")

## End(Not run)

[Package brms version 2.3.0 Index]