print.brmsfit {brms}	R Documentation

Print a summary for a fitted model represented by a brmsfit object

Description

Print a summary for a fitted model represented by a brmsfit object

Usage

## S3 method for class 'brmsfit'
print(x, digits = 2, ...)

Arguments

x
An object of class brmsfit

digits
The number of significant digits for printing out the summary; defaults to 2. The effective sample size is always rounded to integers.

...
Additional arguments that would be passed to method summary of brmsfit.

Author(s)

Paul-Christian Buerkner paul.buerkner@gmail.com

[Package brms version 2.3.0 Index]