prior_summary.brmsfit {brms}	R Documentation

Extract Priors of a Bayesian Model Fitted with brms

Description

Extract Priors of a Bayesian Model Fitted with brms

Usage

## S3 method for class 'brmsfit'
prior_summary(object, all = TRUE, ...)

Arguments

object
A brmsfit object

all
Logical; Show all parameters in the model which may have priors (TRUE) or only those with proper priors (FALSE)?

...
Further arguments passed to or from other methods.

Value

For brmsfit objects, an object of class brmsprior.

Examples

## Not run: 
fit <- brm(count ~ log_Age_c + log_Base4_c * Trt_c  
             + (1|patient) + (1|obs), 
           data = epilepsy, family = poisson(), 
           prior = c(prior(student_t(5,0,10), class = b),
                     prior(cauchy(0,2), class = sd)))
                   
prior_summary(fit)
prior_summary(fit, all = FALSE)
print(prior_summary(fit, all = FALSE), show_df = FALSE)

## End(Not run)

[Package brms version 2.3.0 Index]