Extract Group-Level Estimates

Extract the group-level ('random') effects of each level from a brmsfit object.

# S3 method for brmsfit
ranef(object, summary = TRUE, robust = FALSE,
  probs = c(0.025, 0.975), ...)

Arguments

object
An object of class brmsfit.

summary
Should summary statistics (i.e. means, sds, and 95% intervals) be returned instead of the raw values? Default is TRUE.

robust
If FALSE (the default) the mean is used as the measure of central tendency and the standard deviation as the measure of variability. If TRUE, the median and the median absolute deviation (MAD) are applied instead. Only used if summary is TRUE.

probs
The percentiles to be computed by the quantile function. Only used if summary is TRUE.

...
Currently ignored.

Value

If old is FALSE: A list of arrays (one per grouping factor). If summary is TRUE, names of the first dimension are the factor levels and names of the third dimension are the group-level effects. If summary is FALSE, names of the second dimension are the factor levels and names of the third dimension are the group-level effects.

Examples

# NOT RUN {
fit <- brm(count ~ log_Age_c + log_Base4_c * Trt_c + (1+Trt_c|visit),
           data = epilepsy, family = gaussian(), chains = 2)
ranef(fit)
# }