rows2labels {brms}	R Documentation

Convert Rows to Labels

Description

Convert information in rows to labels for each row.

Usage

rows2labels(x, digits = 2, sep = " & ", incl_vars = TRUE, ...)

Arguments

x
A data.frame for which to extract labels.

digits
Minimal number of decimal places shown in the labels of numeric variables.

sep
A single character string defining the separator between variables used in the labels.

incl_vars
Indicates if variable names should be part of the labels. Defaults to TRUE.

...
Currently unused.

Value

A character vector of the same length as the number of rows of x.

See Also

make_conditions, codemarginal_effects

[Package brms version 2.3.0 Index]