s {brms}	R Documentation

Defining smooths in brms formulas

Description

Functions used in definition of smooth terms within a model formulas. The function does not evaluate a (spline) smooth - it exists purely to help set up a model using spline based smooths.

Usage

s(...)

t2(...)

Arguments

...
Arguments passed to mgcv::s or mgcv::t2.

Details

The function defined here are just simple wrappers of the respective functions of the mgcv package.

See Also

brmsformula, mgcv::s, mgcv::t2

Examples

## Not run: 
# simulate some data
dat <- mgcv::gamSim(1, n = 200, scale = 2)

# fit univariate smooths for all predictors
fit1 <- brm(y ~ s(x0) + s(x1) + s(x2) + s(x3), 
            data = dat, chains = 2)
summary(fit1)
plot(marginal_smooths(fit1), ask = FALSE)

# fit a more complicated smooth model
fit2 <- brm(y ~ t2(x0, x1) + s(x2, by = x3), 
            data = dat, chains = 2)
summary(fit2)
plot(marginal_smooths(fit2), ask = FALSE)

## End(Not run)

[Package brms version 2.3.0 Index]