standata.brmsfit {brms}	R Documentation

Extract Data passed to Stan

Description

Extract all data that was used by Stan to fit the model

Usage

## S3 method for class 'brmsfit'
standata(object, newdata = NULL, re_formula = NULL,
  incl_autocor = TRUE, new_objects = list(), internal = FALSE,
  control = list(), ...)

standata(object, ...)

Arguments

object
An object of class brmsfit.

newdata
An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used.

re_formula
formula containing group-level effects to be considered in the prediction. If NULL (default), include all group-level effects; if NA, include no group-level effects.

incl_autocor
A flag indicating if ARMA autocorrelation parameters should be included in the predictions. Defaults to TRUE. Setting it to FALSE will not affect other correlation structures such as cor_bsts, or cor_fixed.

new_objects
A named list of objects containing new data, which cannot be passed via argument newdata. Currently, only required for objects passed to cor_sar and cor_fixed.

internal
Logical, indicates if the data should be prepared for internal use in other post-processing methods.

control
A named list currently for internal usage only.

...
More arguments passed to make_standata.

Value

A named list containing the data originally passed to Stan.

[Package brms version 2.3.0 Index]