summary.brmsfit {brms}	R Documentation

Create a summary of a fitted model represented by a brmsfit object

Description

Create a summary of a fitted model represented by a brmsfit object

Usage

## S3 method for class 'brmsfit'
summary(object, priors = FALSE, prob = 0.95,
  mc_se = FALSE, use_cache = TRUE, ...)

Arguments

object
An object of class brmsfit

priors
Logical; Indicating if priors should be included in the summary. Default is FALSE.

prob
A value between 0 and 1 indicating the desired probability to be covered by the uncertainty intervals. The default is 0.95.

mc_se
Logical; Indicating if the uncertainty caused by the MCMC sampling should be shown in the summary. Defaults to FALSE.

use_cache
Logical; Indicating if summary results should be cached for future use by rstan. Defaults to TRUE. For models fitted with earlier versions of brms, it may be necessary to set use_cache to FALSE in order to get the summary method working correctly.

...
Other potential arguments

Author(s)

Paul-Christian Buerkner paul.buerkner@gmail.com

[Package brms version 2.3.0 Index]