Projection Predictive Variable Selection

Perform projection predictive variable selection with the projpred package. See varsel and cv_varsel for more details.

# S3 method for brmsfit
varsel(object, ...)

# S3 method for brmsfit
cv_varsel(object, ...)

Arguments

object
A brmsfit object.

...
Further arguments passed to get_refmodel.brmsfit as well as varsel.refmodel or cv_varsel.refmodel.

Value

A vsel object for which several methods are available in the projpred package.

Examples

if (FALSE) {
# fit a simple model
fit <- brm(count ~ zAge + zBase * Trt,
           data = epilepsy, family = poisson())
summary(fit)

# perform variable selection without cross-validation
vs <- varsel(fit)
summary(vs)
plot(vs)

# perform variable selection with cross-validation
cv_vs <- cv_varsel(fit)
summary(cv_vs)
plot(cv_vs)
}