addition-terms {brms}	R Documentation

Additional Response Information

Description

Provide additional information on the response variable in brms models, such as censoring, truncation, or known measurement error.

Usage

resp_se(x, sigma = FALSE)

resp_weights(x, scale = FALSE)

resp_trials(x)

resp_cat(x)

resp_dec(x)

resp_cens(x, y2 = NULL)

resp_trunc(lb = -Inf, ub = Inf)

resp_mi(sdy = NULL)

Arguments

x
A vector; usually a variable defined in the data. Allowed values depend on the function: resp_se and resp_weights require positive numeric values. resp_trials and resp_cat require positive integers. resp_dec requires 0 and 1, or alternatively 'lower' and 'upper'; resp_cens requires 'left', 'none', 'right', and 'interval' (or equivalently -1, 0, 1, and 2) to indicate left, no, right, or interval censoring.

sigma
Logical; Indicates whether the residual standard deviation parameter sigma should be included in addition to the known measurement error. Defaults to FALSE for backwards compatibility, but setting it to TRUE is usually the better choice.

scale
Logical; Indicates whether weights should be scaled so that the average weight equals one. Defaults to FALSE.

y2
A vector specifying the upper bounds in interval censoring.

lb
A numeric vector or single numeric value specifying the lower truncation bound.

ub
A numeric vector or single numeric value specifying the upper truncation bound.

sdy
Optional known measurement error of the response treated as standard deviation. If specified, handles measurement error and (completely) missing values at the same time using the plausible-values-technique.

Details

These functions are almost solely useful when called in formulas passed to the brms package. Within formulas, the resp_ prefix may be omitted. More information is given in the 'Details' section of brmsformula.

Value

A vector containing additional information on the response variable in an appropriate format.

See Also

brm, brmsformula

Examples

## Not run: 
## Random effects meta-analysis
nstudies <- 20
true_effects <- rnorm(nstudies, 0.5, 0.2)
sei <- runif(nstudies, 0.05, 0.3)
outcomes <- rnorm(nstudies, true_effects, sei)
data1 <- data.frame(outcomes, sei)
fit1 <- brm(outcomes | se(sei, sigma = TRUE) ~ 1,
            data = data1)
summary(fit1)

## Probit regression using the binomial family
n <- sample(1:10, 100, TRUE)  # number of trials
success <- rbinom(100, size = n, prob = 0.4)
x <- rnorm(100)
data2 <- data.frame(n, success, x)
fit2 <- brm(success | trials(n) ~ x, data = data2,
            family = binomial("probit"))
summary(fit2)

## Survival regression modeling the time between the first 
## and second recurrence of an infection in kidney patients.
fit3 <- brm(time | cens(censored) ~ age * sex + disease + (1|patient), 
            data = kidney, family = lognormal())
summary(fit3)

## Poisson model with truncated counts  
fit4 <- brm(count | trunc(ub = 104) ~ log_Base4_c * Trt_c, 
            data = epilepsy, family = poisson())
summary(fit4)

## End(Not run)
  

[Package brms version 2.3.0 Index]