Autocorrelation structures

Specify autocorrelation terms in brms models. Currently supported terms are arma, ar, ma, cosy, unstr, sar, car, and fcor. Terms can be directly specified within the formula, or passed to the autocor argument of brmsformula in the form of a one-sided formula. For deprecated ways of specifying autocorrelation terms, see cor_brms.

Details

The autocor term functions are almost solely useful when called in formulas passed to the brms package. They do not evaluate its arguments – but exist purely to help set up a model with autocorrelation terms.

See also

Examples

# specify autocor terms within the formula
y ~ x + arma(p = 1, q = 1) + car(M)
#> y ~ x + arma(p = 1, q = 1) + car(M)
#> <environment: 0x1230da828>

# specify autocor terms in the 'autocor' argument
bf(y ~ x, autocor = ~ arma(p = 1, q = 1) + car(M))
#> y ~ x 
#> autocor ~ arma(p = 1, q = 1) + car(M)

# specify autocor terms via 'acformula'
bf(y ~ x) + acformula(~ arma(p = 1, q = 1) + car(M))
#> y ~ x 
#> autocor ~ arma(p = 1, q = 1) + car(M)