Descriptions of brmshypothesis Objects

A brmshypothesis object contains posterior samples as well as summary statistics of non-linear hypotheses as returned by hypothesis.

# S3 method for brmshypothesis
print(x, digits = 2, chars = 20, ...)

# S3 method for brmshypothesis
plot(
  x,
  N = 5,
  ignore_prior = FALSE,
  chars = 40,
  colors = NULL,
  theme = NULL,
  ask = TRUE,
  plot = TRUE,
  ...
)

Arguments

x
An object of class brmsfit.

digits
Minimal number of significant digits, see print.default.

chars
Maximum number of characters of each hypothesis to print or plot. If NULL, print the full hypotheses. Defaults to 20.

...
Currently ignored.

N
The number of parameters plotted per page.

ignore_prior
A flag indicating if prior distributions should also be plotted. Only used if priors were specified on the relevant parameters.

colors
Two values specifying the colors of the posterior and prior density respectively. If NULL (the default) colors are taken from the current color scheme of the bayesplot package.

theme
A theme object modifying the appearance of the plots. For some basic themes see ggtheme and theme_default.

ask
Logical; indicates if the user is prompted before a new page is plotted. Only used if plot is TRUE.

plot
Logical; indicates if plots should be plotted directly in the active graphic device. Defaults to TRUE.

Details

The two most important elements of a brmshypothesis object are hypothesis, which is a data.frame containing the summary estimates of the hypotheses, and samples, which is a data.frame containing the corresponding posterior samples.

See also