(Deprecated) Spatial simultaneous autoregressive (SAR) structures

Thse functions are deprecated. Please see sar for the new syntax. These functions are constructors for the cor_sar class implementing spatial simultaneous autoregressive structures. The lagsar structure implements SAR of the response values: $$y = \rho W y + \eta + e$$ The errorsar structure implements SAR of the residuals: $$y = \eta + u, u = \rho W u + e$$ In the above equations, \(\eta\) is the predictor term and \(e\) are independent normally or t-distributed residuals.

cor_sar(W, type = c("lag", "error"))

cor_lagsar(W)

cor_errorsar(W)

Arguments

W
An object specifying the spatial weighting matrix. Can be either the spatial weight matrix itself or an object of class listw or nb, from which the spatial weighting matrix can be computed.

type
Type of the SAR structure. Either "lag" (for SAR of the response values) or "error" (for SAR of the residuals).

Value

An object of class cor_sar to be used in calls to brm.

Details

Currently, only families gaussian and student support SAR structures.

Examples

if (FALSE) {
data(oldcol, package = "spdep")
fit1 <- brm(CRIME ~ INC + HOVAL, data = COL.OLD,
            autocor = cor_lagsar(COL.nb),
            chains = 2, cores = 2)
summary(fit1)
plot(fit1)

fit2 <- brm(CRIME ~ INC + HOVAL, data = COL.OLD,
            autocor = cor_errorsar(COL.nb),
            chains = 2, cores = 2)
summary(fit2)
plot(fit2)
}