Extract Diagnostic Quantities of brms Models

Extract quantities that can be used to diagnose sampling behavior of the algorithms applied by Stan at the back-end of brms.

# S3 method for brmsfit
log_posterior(object, ...)

# S3 method for brmsfit
nuts_params(object, pars = NULL, ...)

# S3 method for brmsfit
rhat(object, pars = NULL, ...)

# S3 method for brmsfit
neff_ratio(object, pars = NULL, ...)

Arguments

object
A brmsfit object.

...
Arguments passed to individual methods.

pars
An optional character vector of parameter names. For nuts_params these will be NUTS sampler parameter names rather than model parameters. If pars is omitted all parameters are included.

Value

The exact form of the output depends on the method.

Details

For more details see bayesplot-extractors.

Examples

# NOT RUN {
fit <- brm(time ~ age * sex, data = kidney)

lp <- log_posterior(fit)
head(lp)

np <- nuts_params(fit)
str(np)
# extract the number of divergence transitions
sum(subset(np, Parameter == "divergent__")$Value)

head(rhat(fit))
head(neff_ratio(fit))
# }