Epileptic seizure counts

Breslow and Clayton (1993) analyze data initially provided by Thall and Vail (1990) concerning seizure counts in a randomized trial of anti-convulsant therapy in epilepsy. Covariates are treatment, 8-week baseline seizure counts, and age of the patients in years.

epilepsy

Format

A data frame of 236 observations containing information on the following 9 variables.

Age

The age of the patients in years

Base

The seizure count at 8-weeks baseline

Trt

Either 0 or 1 indicating if the patient received anti-convulsant therapy

patient

The patient number

visit

The session number from 1 (first visit) to 4 (last visit)

count

The seizure count between two visits

obs

The observation number, that is a unique identifier for each observation

zAge

Standardized Age

zBase

Standardized Base

Source

Thall, P. F., & Vail, S. C. (1990). Some covariance models for longitudinal count data with overdispersion. Biometrics, 46(2), 657-671.

Breslow, N. E., & Clayton, D. G. (1993). Approximate inference in generalized linear mixed models. Journal of the American Statistical Association, 88(421), 9-25.

Examples

if (FALSE) {
## poisson regression without random effects.
fit1 <- brm(count ~ zAge + zBase * Trt,
            data = epilepsy, family = poisson())
summary(fit1)
plot(fit1)

## poisson regression with varying intercepts of patients
## as well as normal priors for overall effects parameters.
fit2 <- brm(count ~ zAge + zBase * Trt + (1|patient),
            data = epilepsy, family = poisson(),
            prior = set_prior("normal(0,5)"))
summary(fit2)
plot(fit2)
}