Extract Population-Level Estimates

Extract the population-level ('fixed') effects from a brmsfit object.

# S3 method for brmsfit
fixef(object, summary = TRUE, robust = FALSE,
  probs = c(0.025, 0.975), pars = NULL, ...)

Arguments

object
An object of class brmsfit.

summary
Should summary statistics (i.e. means, sds, and 95% intervals) be returned instead of the raw values? Default is TRUE.

robust
If FALSE (the default) the mean is used as the measure of central tendency and the standard deviation as the measure of variability. If TRUE, the median and the median absolute deviation (MAD) are applied instead. Only used if summary is TRUE.

probs
The percentiles to be computed by the quantile function. Only used if summary is TRUE.

pars
Optional names of coefficients to extract. By default, all coefficients are extracted.

...
Currently ignored.

Value

If summary is TRUE, a matrix with one row per population-level effect and one column per calculated estimate. If summary is FALSE, a matrix with one row per posterior sample and one column per population-level effect.

Examples

# NOT RUN {
fit <- brm(time | cens(censored) ~ age + sex + disease,
           data = kidney, family = "exponential")
fixef(fit)
# extract only some coefficients
fixef(fit, pars = c("age", "sex"))
# }