Overview on Priors for brms Models

Get information on all parameters (and parameter classes) for which priors may be specified including default priors.

get_prior(formula, data, family = gaussian(), autocor = NULL,
  internal = FALSE)

Arguments

formula
An object of class formula, brmsformula, or mvbrmsformula (or one that can be coerced to that classes): A symbolic description of the model to be fitted. The details of model specification are explained in brmsformula.

data
An object of class data.frame (or one that can be coerced to that class) containing data of all variables used in the model.

family
A description of the response distribution and link function to be used in the model. This can be a family function, a call to a family function or a character string naming the family. Every family function has a link argument allowing to specify the link function to be applied on the response variable. If not specified, default links are used. For details of supported families see brmsfamily. By default, a linear gaussian model is applied. In multivariate models, family might also be a list of families.

autocor
An optional cor_brms object describing the correlation structure within the response variable (i.e., the 'autocorrelation'). See the documentation of cor_brms for a description of the available correlation structures. Defaults to NULL, corresponding to no correlations. In multivariate models, autocor might also be a list of autocorrelation structures.

internal
A flag indicating if the names of additional internal parameters should be displayed. Setting priors on these parameters is not recommended

Value

A data.frame with columns prior, class, coef, and group and several rows, each providing information on a parameter (or parameter class) on which priors can be specified. The prior column is empty except for internal default priors.

See also

set_prior

Examples

## get all parameters and parameters classes to define priors on
(prior <- get_prior(count ~ log_Age_c + log_Base4_c * Trt_c
                    + (1|patient) + (1|visit),
                    data = epilepsy, family = poisson()))

#>                  prior     class              coef   group resp dpar nlpar
#> 1                              b                                          
#> 2                              b         log_Age_c                        
#> 3                              b       log_Base4_c                        
#> 4                              b log_Base4_c:Trt_c                        
#> 5                              b             Trt_c                        
#> 6  student_t(3, 1, 10) Intercept                                          
#> 7  student_t(3, 0, 10)        sd                                          
#> 8                             sd                   patient                
#> 9                             sd         Intercept patient                
#> 10                            sd                     visit                
#> 11                            sd         Intercept   visit                
#>    bound
#> 1       
#> 2       
#> 3       
#> 4       
#> 5       
#> 6       
#> 7       
#> 8       
#> 9       
#> 10      
#> 11      

## define a prior on all population-level effects a once
prior$prior[1] <- "normal(0,10)"

## define a specific prior on the population-level effect of Trt_c
prior$prior[5] <- "student_t(10, 0, 5)"

## verify that the priors indeed found their way into Stan's model code
make_stancode(count ~ log_Age_c + log_Base4_c * Trt_c
              + (1|patient) + (1|visit),
              data = epilepsy, family = poisson(),
              prior = prior)

#> // generated with brms 2.2.1
#> functions { 
#> } 
#> data { 
#>   int<lower=1> N;  // total number of observations 
#>   int Y[N];  // response variable 
#>   int<lower=1> K;  // number of population-level effects 
#>   matrix[N, K] X;  // population-level design matrix 
#>   // data for group-level effects of ID 1
#>   int<lower=1> J_1[N];
#>   int<lower=1> N_1;
#>   int<lower=1> M_1;
#>   vector[N] Z_1_1;
#>   // data for group-level effects of ID 2
#>   int<lower=1> J_2[N];
#>   int<lower=1> N_2;
#>   int<lower=1> M_2;
#>   vector[N] Z_2_1;
#>   int prior_only;  // should the likelihood be ignored? 
#> } 
#> transformed data { 
#>   int Kc = K - 1; 
#>   matrix[N, K - 1] Xc;  // centered version of X 
#>   vector[K - 1] means_X;  // column means of X before centering 
#>   for (i in 2:K) { 
#>     means_X[i - 1] = mean(X[, i]); 
#>     Xc[, i - 1] = X[, i] - means_X[i - 1]; 
#>   } 
#> } 
#> parameters { 
#>   vector[Kc] b;  // population-level effects 
#>   real temp_Intercept;  // temporary intercept 
#>   vector<lower=0>[M_1] sd_1;  // group-level standard deviations
#>   vector[N_1] z_1[M_1];  // unscaled group-level effects
#>   vector<lower=0>[M_2] sd_2;  // group-level standard deviations
#>   vector[N_2] z_2[M_2];  // unscaled group-level effects
#> } 
#> transformed parameters { 
#>   // group-level effects 
#>   vector[N_1] r_1_1 = sd_1[1] * (z_1[1]);
#>   // group-level effects 
#>   vector[N_2] r_2_1 = sd_2[1] * (z_2[1]);
#> } 
#> model { 
#>   vector[N] mu = Xc * b + temp_Intercept; 
#>   for (n in 1:N) { 
#>     mu[n] += r_1_1[J_1[n]] * Z_1_1[n] + r_2_1[J_2[n]] * Z_2_1[n];
#>   } 
#>   // priors including all constants 
#>   target += normal_lpdf(b[1] | 0,10); 
#>   target += normal_lpdf(b[2] | 0,10); 
#>   target += student_t_lpdf(b[3] | 10, 0, 5); 
#>   target += normal_lpdf(b[4] | 0,10); 
#>   target += student_t_lpdf(temp_Intercept | 3, 1, 10); 
#>   target += student_t_lpdf(sd_1 | 3, 0, 10)
#>     - 1 * student_t_lccdf(0 | 3, 0, 10); 
#>   target += normal_lpdf(z_1[1] | 0, 1);
#>   target += student_t_lpdf(sd_2 | 3, 0, 10)
#>     - 1 * student_t_lccdf(0 | 3, 0, 10); 
#>   target += normal_lpdf(z_2[1] | 0, 1);
#>   // likelihood including all constants 
#>   if (!prior_only) { 
#>     target += poisson_log_lpmf(Y | mu); 
#>   } 
#> } 
#> generated quantities { 
#>   // actual population-level intercept 
#>   real b_Intercept = temp_Intercept - dot_product(means_X, b); 
#> }