launch_shinystan.brmsfit {brms}	R Documentation

Interface to shinystan

Description

Provide an interface to shinystan for models fitted with brms

Usage

## S3 method for class 'brmsfit'
launch_shinystan(object,
  rstudio = getOption("shinystan.rstudio"), ...)

Arguments

object
A fitted model object typically of class brmsfit.

rstudio
Only relevant for RStudio users. The default (rstudio=FALSE) is to launch the app in the default web browser rather than RStudio's pop-up Viewer. Users can change the default to TRUE by setting the global option
options(shinystan.rstudio = TRUE).

...
Optional arguments to pass to runApp

Value

An S4 shinystan object

See Also

launch_shinystan

Examples

## Not run: 
fit <- brm(rating ~ treat + period + carry + (1|subject),
           data = inhaler, family = "gaussian")
launch_shinystan(fit)                         

## End(Not run)

[Package brms version 2.3.0 Index]