Interface to shinystan

Provide an interface to shinystan for models fitted with brms

# S3 method for brmsfit
launch_shinystan(object,
  rstudio = getOption("shinystan.rstudio"), ...)

Arguments

object
A fitted model object typically of class brmsfit.

rstudio
Only relevant for RStudio users. The default (rstudio=FALSE) is to launch the app in the default web browser rather than RStudio's pop-up Viewer. Users can change the default to TRUE by setting the global option
options(shinystan.rstudio = TRUE).

...
Optional arguments to pass to runApp

Value

An S4 shinystan object

See also

Examples

# NOT RUN {
fit <- brm(rating ~ treat + period + carry + (1|subject),
           data = inhaler, family = "gaussian")
launch_shinystan(fit)
# }