Prepare Fully Crossed Conditions

This is a helper function to prepare fully crossed conditions primarily for use with the conditions argument of conditional_effects. Automatically creates labels for each row in the cond__ column.

Usage

make_conditions(x, vars, ...)

Arguments

x

An R object from which to extract the variables that should be part of the conditions.

vars

Names of the variables that should be part of the conditions.

...

Arguments passed to rows2labels.

Value

A data.frame where each row indicates a condition.

Details

For factor like variables, all levels are used as conditions. For numeric variables, mean + (-1:1) * SD are used as conditions.

See also

Examples

df <- data.frame(x = c("a", "b"), y = rnorm(10))
make_conditions(df, vars = c("x", "y"))
#>   x           y            cond__
#> 1 a -0.09077833 x = a & y = -0.09
#> 2 a  0.91229234  x = a & y = 0.91
#> 3 a  1.91536302  x = a & y = 1.92
#> 4 b -0.09077833 x = b & y = -0.09
#> 5 b  0.91229234  x = b & y = 0.91
#> 6 b  1.91536302  x = b & y = 1.92