Stan Code for brms Models

Generate Stan code for brms models

make_stancode(formula, data, family = gaussian(), prior = NULL,
  autocor = NULL, cov_ranef = NULL, sparse = FALSE,
  sample_prior = c("no", "yes", "only"), stanvars = NULL,
  stan_funs = NULL, save_model = NULL, silent = FALSE, ...)

Arguments

formula
An object of class formula, brmsformula, or mvbrmsformula (or one that can be coerced to that classes): A symbolic description of the model to be fitted. The details of model specification are explained in brmsformula.

data
An object of class data.frame (or one that can be coerced to that class) containing data of all variables used in the model.

family
A description of the response distribution and link function to be used in the model. This can be a family function, a call to a family function or a character string naming the family. Every family function has a link argument allowing to specify the link function to be applied on the response variable. If not specified, default links are used. For details of supported families see brmsfamily. By default, a linear gaussian model is applied. In multivariate models, family might also be a list of families.

prior
One or more brmsprior objects created by set_prior or related functions and combined using the c method or the + operator. See also get_prior for more help.

autocor
An optional cor_brms object describing the correlation structure within the response variable (i.e., the 'autocorrelation'). See the documentation of cor_brms for a description of the available correlation structures. Defaults to NULL, corresponding to no correlations. In multivariate models, autocor might also be a list of autocorrelation structures.

cov_ranef
A list of matrices that are proportional to the (within) covariance structure of the group-level effects. The names of the matrices should correspond to columns in data that are used as grouping factors. All levels of the grouping factor should appear as rownames of the corresponding matrix. This argument can be used, among others to model pedigrees and phylogenetic effects. See vignette("brms_phylogenetics") for more details.

sparse
Logical; indicates whether the population-level design matrices should be treated as sparse (defaults to FALSE). For design matrices with many zeros, this can considerably reduce required memory. Sampling speed is currently not improved or even slightly decreased.

sample_prior
Indicate if samples from all specified proper priors should be drawn additionally to the posterior samples (defaults to "no"). Among others, these samples can be used to calculate Bayes factors for point hypotheses. If set to "only", samples are drawn solely from the priors ignoring the likelihood. In this case, all parameters must have proper priors.

stanvars
An optional stanvars object generated by function stanvar to define additional variables in the data block of Stan.

stan_funs
An optional character string containing self-defined Stan functions, which will be included in the functions block of the generated Stan code.

save_model
Either NULL or a character string. In the latter case, the model code is saved in a file named after the string supplied in save_model, which may also contain the full path where to save the file. If only a name is given, the file is saved in the current working directory.

silent
logical; If TRUE, warnings of the Stan parser will be suppressed.

...
Other arguments for internal usage only

Value

A character string containing the fully commented Stan code to fit a brms model.

Examples

make_stancode(rating ~ treat + period + carry + (1|subject),
              data = inhaler, family = "cumulative")

#> // generated with brms 2.3.1
#> functions { 
#>   /* cumulative-logit log-PDF for a single response
#>    * Args:
#>    *   y: response category
#>    *   mu: linear predictor
#>    *   thres: ordinal thresholds
#>    *   disc: discrimination parameter
#>    * Returns:
#>    *   a scalar to be added to the log posterior
#>    */
#>    real cumulative_logit_lpmf(int y, real mu, vector thres, real disc) {
#>      int ncat = num_elements(thres) + 1;
#>      real p;
#>      if (y == 1) {
#>        p = inv_logit(disc * (thres[1] - mu));
#>      } else if (y == ncat) {
#>        p = 1 - inv_logit(disc * (thres[ncat - 1] - mu));
#>      } else {
#>        p = inv_logit(disc * (thres[y] - mu)) -
#>            inv_logit(disc * (thres[y - 1] - mu));
#>      }
#>      return log(p);
#>    } 
#>   /* cumulative-logit log-PDF for a single response
#>    * including category specific effects
#>    * Args:
#>    *   y: response category
#>    *   mu: linear predictor
#>    *   mucs: predictor for category specific effects
#>    *   thres: ordinal thresholds
#>    *   disc: discrimination parameter
#>    * Returns:
#>    *   a scalar to be added to the log posterior
#>    */
#>    real cumulative_logit_cs_lpmf(int y, real mu, row_vector mucs, vector thres, real disc) {
#>      int ncat = num_elements(thres) + 1;
#>      real p;
#>      if (y == 1) {
#>        p = inv_logit(disc * (thres[1] - mucs[1] - mu));
#>      } else if (y == ncat) {
#>        p = 1 - inv_logit(disc * (thres[ncat - 1] - mucs[ncat - 1] - mu));
#>      } else {
#>        p = inv_logit(disc * (thres[y] - mucs[y] - mu)) -
#>            inv_logit(disc * (thres[y - 1] - mucs[y - 1] - mu));
#>      }
#>      return log(p);
#>    } 
#> } 
#> data { 
#>   int<lower=1> N;  // total number of observations 
#>   int Y[N];  // response variable 
#>   int<lower=2> ncat;  // number of categories 
#>   int<lower=1> K;  // number of population-level effects 
#>   matrix[N, K] X;  // population-level design matrix 
#>   real<lower=0> disc;  // discrimination parameters 
#>   // data for group-level effects of ID 1
#>   int<lower=1> J_1[N];
#>   int<lower=1> N_1;
#>   int<lower=1> M_1;
#>   vector[N] Z_1_1;
#>   int prior_only;  // should the likelihood be ignored? 
#> } 
#> transformed data { 
#>   int Kc = K - 1; 
#>   matrix[N, K - 1] Xc;  // centered version of X 
#>   vector[K - 1] means_X;  // column means of X before centering 
#>   for (i in 2:K) { 
#>     means_X[i - 1] = mean(X[, i]); 
#>     Xc[, i - 1] = X[, i] - means_X[i - 1]; 
#>   } 
#> } 
#> parameters { 
#>   vector[Kc] b;  // population-level effects 
#>   ordered[ncat-1] temp_Intercept;  // temporary thresholds 
#>   vector<lower=0>[M_1] sd_1;  // group-level standard deviations
#>   vector[N_1] z_1[M_1];  // unscaled group-level effects
#> } 
#> transformed parameters { 
#>   // group-level effects 
#>   vector[N_1] r_1_1 = sd_1[1] * (z_1[1]);
#> } 
#> model { 
#>   vector[N] mu = Xc * b; 
#>   for (n in 1:N) { 
#>     mu[n] += r_1_1[J_1[n]] * Z_1_1[n];
#>   } 
#>   // priors including all constants 
#>   target += student_t_lpdf(temp_Intercept | 3, 0, 10); 
#>   target += student_t_lpdf(sd_1 | 3, 0, 10)
#>     - 1 * student_t_lccdf(0 | 3, 0, 10); 
#>   target += normal_lpdf(z_1[1] | 0, 1);
#>   // likelihood including all constants 
#>   if (!prior_only) { 
#>     for (n in 1:N) { 
#>       target += ordered_logistic_lpmf(Y[n] | mu[n], temp_Intercept); 
#>     } 
#>   } 
#> } 
#> generated quantities { 
#>   // compute actual thresholds 
#>   vector[ncat - 1] b_Intercept = temp_Intercept + dot_product(means_X, b); 
#> } 

make_stancode(count ~ log_Age_c + log_Base4_c * Trt_c
              + (1|patient) + (1|visit),
              data = epilepsy, family = "poisson")

#> // generated with brms 2.3.1
#> functions { 
#> } 
#> data { 
#>   int<lower=1> N;  // total number of observations 
#>   int Y[N];  // response variable 
#>   int<lower=1> K;  // number of population-level effects 
#>   matrix[N, K] X;  // population-level design matrix 
#>   // data for group-level effects of ID 1
#>   int<lower=1> J_1[N];
#>   int<lower=1> N_1;
#>   int<lower=1> M_1;
#>   vector[N] Z_1_1;
#>   // data for group-level effects of ID 2
#>   int<lower=1> J_2[N];
#>   int<lower=1> N_2;
#>   int<lower=1> M_2;
#>   vector[N] Z_2_1;
#>   int prior_only;  // should the likelihood be ignored? 
#> } 
#> transformed data { 
#>   int Kc = K - 1; 
#>   matrix[N, K - 1] Xc;  // centered version of X 
#>   vector[K - 1] means_X;  // column means of X before centering 
#>   for (i in 2:K) { 
#>     means_X[i - 1] = mean(X[, i]); 
#>     Xc[, i - 1] = X[, i] - means_X[i - 1]; 
#>   } 
#> } 
#> parameters { 
#>   vector[Kc] b;  // population-level effects 
#>   real temp_Intercept;  // temporary intercept 
#>   vector<lower=0>[M_1] sd_1;  // group-level standard deviations
#>   vector[N_1] z_1[M_1];  // unscaled group-level effects
#>   vector<lower=0>[M_2] sd_2;  // group-level standard deviations
#>   vector[N_2] z_2[M_2];  // unscaled group-level effects
#> } 
#> transformed parameters { 
#>   // group-level effects 
#>   vector[N_1] r_1_1 = sd_1[1] * (z_1[1]);
#>   // group-level effects 
#>   vector[N_2] r_2_1 = sd_2[1] * (z_2[1]);
#> } 
#> model { 
#>   vector[N] mu = Xc * b + temp_Intercept; 
#>   for (n in 1:N) { 
#>     mu[n] += r_1_1[J_1[n]] * Z_1_1[n] + r_2_1[J_2[n]] * Z_2_1[n];
#>   } 
#>   // priors including all constants 
#>   target += student_t_lpdf(temp_Intercept | 3, 1, 10); 
#>   target += student_t_lpdf(sd_1 | 3, 0, 10)
#>     - 1 * student_t_lccdf(0 | 3, 0, 10); 
#>   target += normal_lpdf(z_1[1] | 0, 1);
#>   target += student_t_lpdf(sd_2 | 3, 0, 10)
#>     - 1 * student_t_lccdf(0 | 3, 0, 10); 
#>   target += normal_lpdf(z_2[1] | 0, 1);
#>   // likelihood including all constants 
#>   if (!prior_only) { 
#>     target += poisson_log_lpmf(Y | mu); 
#>   } 
#> } 
#> generated quantities { 
#>   // actual population-level intercept 
#>   real b_Intercept = temp_Intercept - dot_product(means_X, b); 
#> }