me {brms}	R Documentation

Predictors with Measurement Error in brms Models

Description

Specify predictors with measurement error. This function is almost solely useful when called in formulas passed to the brms package.

Usage

me(x, sdx = NULL, gr = NULL)

Arguments

x
The variable measured with error.

sdx
Known measurement error of x treated as standard deviation.

gr
Optional grouping factor to specify which values of x correspond to the same value of the latent variable. If NULL (the default) each observation will have its own value of the latent variable.

Details

For detailed documentation see help(brmsformula).

By default, latent noise-free variables are assumed to be correlated. To change that, add set_mecor(FALSE) to your model formula object (see examples).

See Also

brmsformula, brmsformula-helpers

Examples

## Not run: 
# sample some data
N <- 100
dat <- data.frame(
  y = rnorm(N), x1 = rnorm(N), 
  x2 = rnorm(N), sdx = abs(rnorm(N, 1))
 )
# fit a simple error-in-variables model 
fit1 <- brm(y ~ me(x1, sdx) + me(x2, sdx), data = dat, 
           save_mevars = TRUE)
summary(fit1)

# turn off modeling of correlations
bform <- bf(y ~ me(x1, sdx) + me(x2, sdx)) + set_mecor(FALSE)
fit2 <- brm(bform, data = dat, save_mevars = TRUE)
summary(fit2)

## End(Not run) 

[Package brms version 2.3.0 Index]