Create a matrix of output plots from a brmsfit object

A pairs method that is customized for MCMC output.

# S3 method for brmsfit
pairs(x, pars = NA, fixed = FALSE, exact_match = FALSE, ...)

Arguments

x
An object of class brmsfit

pars
Names of the parameters to plot, as given by a character vector or a regular expression. By default, all parameters except for group-level and smooth effects are plotted.

fixed
Indicates whether parameter names should be matched exactly (TRUE) or treated as regular expressions (FALSE). Default is FALSE.

exact_match
Deprecated alias of argument fixed.

...
Further arguments to be passed to mcmc_pairs.

Details

For a detailed description see mcmc_pairs.

Examples

if (FALSE) {
fit <- brm(count ~ zAge + zBase * Trt 
           + (1|patient) + (1|visit), 
           data = epilepsy, family = "poisson")  
pairs(fit, pars = parnames(fit)[1:3], fixed = TRUE)
pairs(fit, pars = "^sd_")
}