Create a matrix of output plots from a brmsfit object

A pairs method that is customized for MCMC output.

Usage

# S3 method for class 'brmsfit'
pairs(x, pars = NA, variable = NULL, regex = FALSE, fixed = FALSE, ...)

Arguments

x

An object of class brmsfit

pars

Deprecated alias of variable. Names of the parameters to plot, as given by a character vector or a regular expression.

variable

Names of the variables (parameters) to plot, as given by a character vector or a regular expression (if regex = TRUE). By default, a hopefully not too large selection of variables is plotted.

regex

Logical; Indicates whether variable should be treated as regular expressions. Defaults to FALSE.

fixed

(Deprecated) Indicates whether parameter names should be matched exactly (TRUE) or treated as regular expressions (FALSE). Default is FALSE and only works with argument pars.

...

Further arguments to be passed to mcmc_pairs.

Details

For a detailed description see mcmc_pairs.

Examples

if (FALSE) { # \dontrun{
fit <- brm(count ~ zAge + zBase * Trt
           + (1|patient) + (1|visit),
           data = epilepsy, family = "poisson")
pairs(fit, variable = variables(fit)[1:3])
pairs(fit, variable = "^sd_", regex = TRUE)
} # }