Trace and Density Plots for MCMC Draws

Trace and Density Plots for MCMC Draws

Usage

# S3 method for class 'brmsfit'
plot(
  x,
  pars = NA,
  combo = c("hist", "trace"),
  nvariables = 5,
  N = NULL,
  variable = NULL,
  regex = FALSE,
  fixed = FALSE,
  bins = 30,
  theme = NULL,
  plot = TRUE,
  ask = TRUE,
  newpage = TRUE,
  ...
)

Arguments

x

An object of class brmsfit.

pars

Deprecated alias of variable. Names of the parameters to plot, as given by a character vector or a regular expression.

combo

A character vector with at least two elements. Each element of combo corresponds to a column in the resulting graphic and should be the name of one of the available MCMC functions (omitting the mcmc_ prefix).

nvariables

The number of variables (parameters) plotted per page.

N

Deprecated alias of nvariables.

variable

Names of the variables (parameters) to plot, as given by a character vector or a regular expression (if regex = TRUE). By default, a hopefully not too large selection of variables is plotted.

regex

Logical; Indicates whether variable should be treated as regular expressions. Defaults to FALSE.

fixed

(Deprecated) Indicates whether parameter names should be matched exactly (TRUE) or treated as regular expressions (FALSE). Default is FALSE and only works with argument pars.

bins

Number of bins used for posterior histograms (defaults to 30).

theme

A theme object modifying the appearance of the plots. For some basic themes see ggtheme and theme_default.

plot

Logical; indicates if plots should be plotted directly in the active graphic device. Defaults to TRUE.

ask

Logical; indicates if the user is prompted before a new page is plotted. Only used if plot is TRUE.

newpage

Logical; indicates if the first set of plots should be plotted to a new page. Only used if plot is TRUE.

...

Further arguments passed to mcmc_combo.

Value

An invisible list of gtable objects.

Examples

if (FALSE) { # \dontrun{
fit <- brm(count ~ zAge + zBase * Trt
           + (1|patient) + (1|visit),
           data = epilepsy, family = "poisson")
plot(fit)
## plot population-level effects only
plot(fit, variable = "^b_", regex = TRUE)
} # }