Posterior Samples of the Linear Predictor

Compute posterior samples of the linear predictor, This method is an alias of pp_expect.brmsfit with scale = "linear".

# S3 method for brmsfit
posterior_linpred(
  object,
  transform = FALSE,
  newdata = NULL,
  re_formula = NULL,
  re.form = NULL,
  resp = NULL,
  dpar = NULL,
  nlpar = NULL,
  nsamples = NULL,
  subset = NULL,
  sort = FALSE,
  ...
)

Arguments

object
An object of class brmsfit.

transform
Logical; alias of scale. If TRUE, scale is set to "response". If FALSE, scale is set to "linear". Only implemented for compatibility with the posterior_linpred generic.

newdata
An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used. NA values within factors are interpreted as if all dummy variables of this factor are zero. This allows, for instance, to make predictions of the grand mean when using sum coding.

re_formula
formula containing group-level effects to be considered in the prediction. If NULL (default), include all group-level effects; if NA, include no group-level effects.

re.form
Alias of re_formula.

resp
Optional names of response variables. If specified, predictions are performed only for the specified response variables.

dpar
Optional name of a predicted distributional parameter. If specified, fitted values of this parameters are returned.

nlpar
Optional name of a predicted non-linear parameter. If specified, fitted values of this parameters are returned.

nsamples
Positive integer indicating how many posterior samples should be used. If NULL (the default) all samples are used. Ignored if subset is not NULL.

subset
A numeric vector specifying the posterior samples to be used. If NULL (the default), all samples are used.

sort
Logical. Only relevant for time series models. Indicating whether to return predicted values in the original order (FALSE; default) or in the order of the time series (TRUE).

...
Further arguments passed to extract_draws that control several aspects of data validation and prediction.

See also

Examples

if (FALSE) {
## fit a model
fit <- brm(rating ~ treat + period + carry + (1|subject),
           data = inhaler)

## extract linear predictor values
pl <- posterior_linpred(fit)
str(pl)
}