posterior_table {brms}	R Documentation

Table Creation for Posterior Samples

Description

Create a table for unique values of posterior samples. This is usually only useful when summarizing predictions of ordinal models.

Usage

posterior_table(x, levels = NULL)

Arguments

x
A matrix of posterior samples where rows indicate samples and columns indicate parameters.

levels
Optional values of possible posterior values. Defaults to all unique values in x.

Value

A matrix where rows indicate parameters and columns indicate the unique values of posterior samples.

Examples

## Not run: 
fit <- brm(rating ~ period + carry + treat, 
           data = inhaler, family = cumulative())
pr <- predict(fit, summary = FALSE)
posterior_table(pr)

## End(Not run)
 

[Package brms version 2.3.0 Index]