Table Creation for Posterior Draws

Create a table for unique values of posterior draws. This is usually only useful when summarizing predictions of ordinal models.

Usage

posterior_table(x, levels = NULL)

Arguments

x

A matrix of posterior draws where rows indicate draws and columns indicate parameters.

levels

Optional values of possible posterior values. Defaults to all unique values in x.

Value

A matrix where rows indicate parameters and columns indicate the unique values of posterior draws.

Examples

if (FALSE) { # \dontrun{
fit <- brm(rating ~ period + carry + treat,
           data = inhaler, family = cumulative())
pr <- predict(fit, summary = FALSE)
posterior_table(pr)
} # }