Read CmdStan CSV files as a brms-formatted stanfit object

read_csv_as_stanfit is used internally to read CmdStan CSV files into a stanfit object that is consistent with the structure of the fit slot of a brmsfit object.

Usage

read_csv_as_stanfit(
  files,
  variables = NULL,
  sampler_diagnostics = NULL,
  model = NULL,
  exclude = "",
  algorithm = "sampling"
)

Arguments

files

Character vector of CSV files names where draws are stored.

variables

Character vector of variables to extract from the CSV files.

sampler_diagnostics

Character vector of sampler diagnostics to extract.

model

A compiled cmdstanr model object (optional). Provide this argument if you want to allow updating the model without recompilation.

exclude

Character vector of variables to exclude from the stanfit. Only used when variables is also specified.

algorithm

The algorithm with which the model was fitted. See brm for details.

Value

A stanfit object consistent with the structure of the fit slot of a brmsfit object.

Examples

if (FALSE) { # \dontrun{
# fit a model manually via cmdstanr
scode <- stancode(count ~ Trt, data = epilepsy)
sdata <- standata(count ~ Trt, data = epilepsy)
mod <- cmdstanr::cmdstan_model(cmdstanr::write_stan_file(scode))
stanfit <- mod$sample(data = sdata)

# feed the Stan model back into brms
fit <- brm(count ~ Trt, data = epilepsy, empty = TRUE, backend = 'cmdstanr')
fit$fit <- read_csv_as_stanfit(stanfit$output_files(), model = mod)
fit <- rename_pars(fit)
summary(fit)
} # }