Compute exact cross-validation for problematic observations

Compute exact cross-validation for problematic observations for which approximate leave-one-out cross-validation may return incorrect results. Models for problematic observations can be run in parallel using the future package.

reloo(x, fit, k_threshold = 0.7, newdata = NULL, resp = NULL,
  check = TRUE, ...)

Arguments

x
An R object of class loo.

fit
An R object of class brmsfit.

k_threshold
The threshold at which pareto \(k\) estimates are treated as problematic. Defaults to 0.7. See pareto_k_ids for more details.

newdata
An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used.

resp
Optional names of response variables. If specified, predictions are performed only for the specified response variables.

check
Logical; If TRUE (the default), some checks check are performed if the loo object was generated from the brmsfit object passed to argument fit.

...
Further arguments passed to update.brmsfit and log_lik.brmsfit.

Value

An object of the class loo.

Details

Warnings about Pareto \(k\) estimates indicate observations for which the approximation to LOO is problematic (this is described in detail in Vehtari, Gelman, and Gabry (2017) and the loo package documentation). If there are \(J\) observations with \(k\) estimates above k_threshold, then reloo will refit the original model \(J\) times, each time leaving out one of the \(J\) problematic observations. The pointwise contributions of these observations to the total ELPD are then computed directly and substituted for the previous estimates from these \(J\) observations that are stored in the original loo object.

See also

Examples

# NOT RUN {
fit1 <- brm(count ~ zAge + zBase * Trt + (1|patient),
           data = epilepsy, family = poisson())
# throws warning about some pareto k estimates being too high
(loo1 <- loo(fit1))
(reloo1 <- reloo(loo1, fit1, chains = 1))
# }