Extract Model Residuals from brmsfit Objects

Extract Model Residuals from brmsfit Objects

# S3 method for brmsfit
residuals(object, newdata = NULL, re_formula = NULL,
  type = c("ordinary", "pearson"), method = c("fitted", "predict"),
  allow_new_levels = FALSE, sample_new_levels = "uncertainty",
  new_objects = list(), incl_autocor = TRUE, resp = NULL, subset = NULL,
  nsamples = NULL, sort = FALSE, nug = NULL, summary = TRUE,
  robust = FALSE, probs = c(0.025, 0.975), ...)

# S3 method for brmsfit
predictive_error(object, newdata = NULL,
  re_formula = NULL, re.form = NULL, allow_new_levels = FALSE,
  sample_new_levels = "uncertainty", new_objects = list(),
  incl_autocor = TRUE, resp = NULL, subset = NULL, nsamples = NULL,
  sort = FALSE, nug = NULL, robust = FALSE, probs = c(0.025, 0.975),
  ...)

Arguments

object
An object of class brmsfit

newdata
An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used.

re_formula
formula containing group-level effects to be considered in the prediction. If NULL (default), include all group-level effects; if NA, include no group-level effects.

type
The type of the residuals, either "ordinary" or "pearson". More information is provided under 'Details'.

method
Indicates the method to compute model implied values. Either "fitted" (predicted values of the regression curve) or "predict" (predicted response values). Using "predict" is recommended but "fitted" is the current default for reasons of backwards compatibility.

allow_new_levels
A flag indicating if new levels of group-level effects are allowed (defaults to FALSE). Only relevant if newdata is provided.

sample_new_levels
Indicates how to sample new levels for grouping factors specified in re_formula. This argument is only relevant if newdata is provided and allow_new_levels is set to TRUE. If "uncertainty" (default), include group-level uncertainty in the predictions based on the variation of the existing levels. If "gaussian", sample new levels from the (multivariate) normal distribution implied by the group-level standard deviations and correlations. This options may be useful for conducting Bayesian power analysis. If "old_levels", directly sample new levels from the existing levels.

new_objects
A named list of objects containing new data, which cannot be passed via argument newdata. Currently, only required for objects passed to cor_sar and cor_fixed.

incl_autocor
A flag indicating if ARMA autocorrelation parameters should be included in the predictions. Defaults to TRUE. Setting it to FALSE will not affect other correlation structures such as cor_bsts, or cor_fixed.

resp
Optional names of response variables. If specified, fitted values of these response variables are returned.

subset
A numeric vector specifying the posterior samples to be used. If NULL (the default), all samples are used.

nsamples
Positive integer indicating how many posterior samples should be used. If NULL (the default) all samples are used. Ignored if subset is not NULL.

sort
Logical. Only relevant for time series models. Indicating whether to return predicted values in the original order (FALSE; default) or in the order of the time series (TRUE).

nug
Small positive number for Gaussian process terms only. For numerical reasons, the covariance matrix of a Gaussian process might not be positive definite. Adding a very small number to the matrix's diagonal often solves this problem. If NULL (the default), nug is chosen internally.

summary
Should summary statistics (i.e. means, sds, and 95% intervals) be returned instead of the raw values? Default is TRUE.

robust
If FALSE (the default) the mean is used as the measure of central tendency and the standard deviation as the measure of variability. If TRUE, the median and the median absolute deviation (MAD) are applied instead. Only used if summary is TRUE.

probs
The percentiles to be computed by the quantile function. Only used if summary is TRUE.

...
Currently ignored.

re.form
Alias of re_formula.

Value

Model residuals. If summary = TRUE this is a N x C matrix and if summary = FALSE a S x N matrix, where S is the number of samples, N is the number of observations, and C is equal to length(probs) + 2.

Details

Residuals of type ordinary are of the form \(R = Y - Yp\), where \(Y\) is the observed and \(Yp\) is the predicted response. Residuals of type pearson are of the form \(R = (Y - Yp) / SD(Y)\), where \(SD(Y)\) is an estimation of the standard deviation of \(Y\).

Currently, residuals.brmsfit does not support categorical or ordinal models.

Method predictive_error.brmsfit is an alias of residuals.brmsfit with method = "predict" and summary = FALSE.

Examples

# NOT RUN {
## fit a model
fit <- brm(rating ~ treat + period + carry + (1|subject),
           data = inhaler, cores = 2)

## extract residuals 
res <- residuals(fit, summary = TRUE)
head(res)
# }