Posterior Samples of Residuals/Predictive Errors

This method is an alias of predictive_error.brmsfit with additional arguments for obtaining summaries of the computed samples.

# S3 method for brmsfit
residuals(
  object,
  newdata = NULL,
  re_formula = NULL,
  method = "pp_expect",
  type = c("ordinary", "pearson"),
  resp = NULL,
  nsamples = NULL,
  subset = NULL,
  sort = FALSE,
  summary = TRUE,
  robust = FALSE,
  probs = c(0.025, 0.975),
  ...
)

Arguments

object
An object of class brmsfit.

newdata
An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used. NA values within factors are interpreted as if all dummy variables of this factor are zero. This allows, for instance, to make predictions of the grand mean when using sum coding.

re_formula
formula containing group-level effects to be considered in the prediction. If NULL (default), include all group-level effects; if NA, include no group-level effects.

method
Method use to obtain predictions. Either "pp_expect" (the default) or "posterior_predict". Using "posterior_predict" is recommended but "pp_expect" is the current default for reasons of backwards compatibility.

type
The type of the residuals, either "ordinary" or "pearson". More information is provided under 'Details'.

resp
Optional names of response variables. If specified, predictions are performed only for the specified response variables.

nsamples
Positive integer indicating how many posterior samples should be used. If NULL (the default) all samples are used. Ignored if subset is not NULL.

subset
A numeric vector specifying the posterior samples to be used. If NULL (the default), all samples are used.

sort
Logical. Only relevant for time series models. Indicating whether to return predicted values in the original order (FALSE; default) or in the order of the time series (TRUE).

summary
Should summary statistics be returned instead of the raw values? Default is TRUE..

robust
If FALSE (the default) the mean is used as the measure of central tendency and the standard deviation as the measure of variability. If TRUE, the median and the median absolute deviation (MAD) are applied instead. Only used if summary is TRUE.

probs
The percentiles to be computed by the quantile function. Only used if summary is TRUE.

...
Further arguments passed to extract_draws that control several aspects of data validation and prediction.

Value

An array of predictive error/residual samples. If summary = FALSE the output resembles those of predictive_error.brmsfit. If summary = TRUE the output is an N x E matrix, where N is the number of observations and E denotes the summary statistics computed from the samples.

Details

Residuals of type 'ordinary' are of the form \(R = Y - Yrep\), where \(Y\) is the observed and \(Yrep\) is the predicted response. Residuals of type pearson are of the form \(R = (Y - Yrep) / SD(Y)\), where \(SD(Y)\) is an estimation of the standard deviation of \(Y\).

Examples

if (FALSE) {
## fit a model
fit <- brm(rating ~ treat + period + carry + (1|subject),
           data = inhaler, cores = 2)

## extract residuals/predictive errors
res <- residuals(fit)
head(res)
}