Extract Model Residuals from brmsfit Objects

Extract Model Residuals from brmsfit Objects

# S3 method for brmsfit
residuals(object, newdata = NULL, re_formula = NULL,
  type = c("ordinary", "pearson"), method = c("fitted", "predict"),
  resp = NULL, nsamples = NULL, subset = NULL, sort = FALSE,
  summary = TRUE, robust = FALSE, probs = c(0.025, 0.975), ...)

# S3 method for brmsfit
predictive_error(object, newdata = NULL,
  re_formula = NULL, re.form = NULL, resp = NULL, nsamples = NULL,
  subset = NULL, sort = FALSE, robust = FALSE, probs = c(0.025,
  0.975), ...)

Arguments

object
An object of class brmsfit.

newdata
An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used.

re_formula
formula containing group-level effects to be considered in the prediction. If NULL (default), include all group-level effects; if NA, include no group-level effects.

type
The type of the residuals, either "ordinary" or "pearson". More information is provided under 'Details'.

method
Indicates the method to compute model implied values. Either "fitted" (predicted values of the regression curve) or "predict" (predicted response values). Using "predict" is recommended but "fitted" is the current default for reasons of backwards compatibility.

resp
Optional names of response variables. If specified, predictions are performed only for the specified response variables.

nsamples
Positive integer indicating how many posterior samples should be used. If NULL (the default) all samples are used. Ignored if subset is not NULL.

subset
A numeric vector specifying the posterior samples to be used. If NULL (the default), all samples are used.

sort
Logical. Only relevant for time series models. Indicating whether to return predicted values in the original order (FALSE; default) or in the order of the time series (TRUE).

summary
Should summary statistics (i.e. means, sds, and 95% intervals) be returned instead of the raw values? Default is TRUE.

robust
If FALSE (the default) the mean is used as the measure of central tendency and the standard deviation as the measure of variability. If TRUE, the median and the median absolute deviation (MAD) are applied instead. Only used if summary is TRUE.

probs
The percentiles to be computed by the quantile function. Only used if summary is TRUE.

...
Further arguments passed to extract_draws that control several aspects of data validation and prediction.

re.form
Alias of re_formula.

Value

Model residuals. If summary = TRUE this is a N x C matrix and if summary = FALSE a S x N matrix, where S is the number of samples, N is the number of observations, and C is equal to length(probs) + 2.

Details

Residuals of type ordinary are of the form \(R = Y - Yp\), where \(Y\) is the observed and \(Yp\) is the predicted response. Residuals of type pearson are of the form \(R = (Y - Yp) / SD(Y)\), where \(SD(Y)\) is an estimation of the standard deviation of \(Y\).

Currently, residuals.brmsfit does not support categorical or ordinal models.

Method predictive_error.brmsfit is an alias of residuals.brmsfit with method = "predict" and summary = FALSE.

Examples

# NOT RUN {
## fit a model
fit <- brm(rating ~ treat + period + carry + (1|subject),
           data = inhaler, cores = 2)

## extract residuals 
res <- residuals(fit, summary = TRUE)
head(res)
# }