Restructure Old brmsfit Objects

Restructure old brmsfit objects to work with the latest brms version. This function is called internally when applying post-processing methods. However, in order to avoid unnecessary run time caused by the restructuring, I recommend explicitly calling restructure once per model after updating brms.

restructure(x, rstr_summary = FALSE)

Arguments

x
An object of class brmsfit.

rstr_summary
Logical; If TRUE, the cached summary stored by rstan is restructured as well.

Value

A brmsfit object compatible with the latest version of brms.