User-defined variables passed to Stan

Prepare user-defined variables to be passed to Stan's data block. This is primarily useful for defining more complex priors and for refitting models without recompilation despite changing priors.

stanvar(x, name = NULL, scode = NULL)

Arguments

x
An R object containing data to be passed to Stan.

name
Optinal character string providing the desired variable name of the object in x. If NULL (the default) the variable name is directly infered from x.

scode
Optional line of Stan code to define the variable in Stan's data block. If NULL (the default), the Stan code is inferred based on the class of x.

Value

An object of class stanvars.

Examples

bprior <- prior(normal(mean_intercept, 10), class = "Intercept")
stanvars <- stanvar(5, name = "mean_intercept")
make_stancode(count ~ Trt, epilepsy, prior = bprior,
              stanvars = stanvars)

#> // generated with brms 2.3.1
#> functions { 
#> } 
#> data { 
#>   int<lower=1> N;  // total number of observations 
#>   vector[N] Y;  // response variable 
#>   int<lower=1> K;  // number of population-level effects 
#>   matrix[N, K] X;  // population-level design matrix 
#>   int prior_only;  // should the likelihood be ignored? 
#>   real mean_intercept;
#> } 
#> transformed data { 
#>   int Kc = K - 1; 
#>   matrix[N, K - 1] Xc;  // centered version of X 
#>   vector[K - 1] means_X;  // column means of X before centering 
#>   for (i in 2:K) { 
#>     means_X[i - 1] = mean(X[, i]); 
#>     Xc[, i - 1] = X[, i] - means_X[i - 1]; 
#>   } 
#> } 
#> parameters { 
#>   vector[Kc] b;  // population-level effects 
#>   real temp_Intercept;  // temporary intercept 
#>   real<lower=0> sigma;  // residual SD 
#> } 
#> transformed parameters { 
#> } 
#> model { 
#>   vector[N] mu = Xc * b + temp_Intercept; 
#>   // priors including all constants 
#>   target += normal_lpdf(temp_Intercept | mean_intercept, 10); 
#>   target += student_t_lpdf(sigma | 3, 0, 10)
#>     - 1 * student_t_lccdf(0 | 3, 0, 10); 
#>   // likelihood including all constants 
#>   if (!prior_only) { 
#>     target += normal_lpdf(Y | mu, sigma); 
#>   } 
#> } 
#> generated quantities { 
#>   // actual population-level intercept 
#>   real b_Intercept = temp_Intercept - dot_product(means_X, b); 
#> } 

# define a multi-normal prior with known covariance matrix
bprior <- prior(multi_normal(M, V), class = "b")
stanvars <- stanvar(rep(0, 2), "M", scode = "  vector[K] M;") +
  stanvar(diag(2), "V", scode = "  matrix[K, K] V;")
make_stancode(count ~ Trt + log_Base4_c, epilepsy,
              prior = bprior, stanvars = stanvars)

#> // generated with brms 2.3.1
#> functions { 
#> } 
#> data { 
#>   int<lower=1> N;  // total number of observations 
#>   vector[N] Y;  // response variable 
#>   int<lower=1> K;  // number of population-level effects 
#>   matrix[N, K] X;  // population-level design matrix 
#>   int prior_only;  // should the likelihood be ignored? 
#>   vector[K] M;
#>   matrix[K, K] V;
#> } 
#> transformed data { 
#>   int Kc = K - 1; 
#>   matrix[N, K - 1] Xc;  // centered version of X 
#>   vector[K - 1] means_X;  // column means of X before centering 
#>   for (i in 2:K) { 
#>     means_X[i - 1] = mean(X[, i]); 
#>     Xc[, i - 1] = X[, i] - means_X[i - 1]; 
#>   } 
#> } 
#> parameters { 
#>   vector[Kc] b;  // population-level effects 
#>   real temp_Intercept;  // temporary intercept 
#>   real<lower=0> sigma;  // residual SD 
#> } 
#> transformed parameters { 
#> } 
#> model { 
#>   vector[N] mu = Xc * b + temp_Intercept; 
#>   // priors including all constants 
#>   target += multi_normal_lpdf(b | M, V); 
#>   target += student_t_lpdf(temp_Intercept | 3, 4, 10); 
#>   target += student_t_lpdf(sigma | 3, 0, 10)
#>     - 1 * student_t_lccdf(0 | 3, 0, 10); 
#>   // likelihood including all constants 
#>   if (!prior_only) { 
#>     target += normal_lpdf(Y | mu, sigma); 
#>   } 
#> } 
#> generated quantities { 
#>   // actual population-level intercept 
#>   real b_Intercept = temp_Intercept - dot_product(means_X, b); 
#> }