User-defined variables passed to Stan

Prepare user-defined variables to be passed to one of Stan's program blocks. This is primarily useful for defining more complex priors, for refitting models without recompilation despite changing priors, or for defining custom Stan functions.

stanvar(x = NULL, name = NULL, scode = NULL, block = "data")

Arguments

x
An R object containing data to be passed to Stan. Only required if block = 'data' and ignored otherwise.

name
Optinal character string providing the desired variable name of the object in x. If NULL (the default) the variable name is directly infered from x.

scode
Line of Stan code to define the variable in Stan language. If block = 'data', the Stan code is inferred based on the class of x by default.

block
Name of one of Stan's program blocks in which the variable should be defined. Can be 'data', 'tdata' (transformed data), 'parameters', 'tparameters' (transformed parameters), 'model', 'genquant' (generated quantities) or 'functions'.

Value

An object of class stanvars.

Examples

bprior <- prior(normal(mean_intercept, 10), class = "Intercept")
stanvars <- stanvar(5, name = "mean_intercept")
make_stancode(count ~ Trt, epilepsy, prior = bprior,
              stanvars = stanvars)

#> // generated with brms 2.10.3
#> functions {
#> }
#> data {
#>   int<lower=1> N;  // number of observations
#>   vector[N] Y;  // response variable
#>   int<lower=1> K;  // number of population-level effects
#>   matrix[N, K] X;  // population-level design matrix
#>   int prior_only;  // should the likelihood be ignored?
#>   real mean_intercept;
#> }
#> transformed data {
#>   int Kc = K - 1;
#>   matrix[N, Kc] Xc;  // centered version of X without an intercept
#>   vector[Kc] means_X;  // column means of X before centering
#>   for (i in 2:K) {
#>     means_X[i - 1] = mean(X[, i]);
#>     Xc[, i - 1] = X[, i] - means_X[i - 1];
#>   }
#> }
#> parameters {
#>   vector[Kc] b;  // population-level effects
#>   // temporary intercept for centered predictors
#>   real Intercept;
#>   real<lower=0> sigma;  // residual SD
#> }
#> transformed parameters {
#> }
#> model {
#>   // priors including all constants
#>   target += normal_lpdf(Intercept | mean_intercept, 10);
#>   target += student_t_lpdf(sigma | 3, 0, 10)
#>     - 1 * student_t_lccdf(0 | 3, 0, 10);
#>   // likelihood including all constants
#>   if (!prior_only) {
#>     target += normal_id_glm_lpdf(Y | Xc, Intercept, b, sigma);
#>   }
#> }
#> generated quantities {
#>   // actual population-level intercept
#>   real b_Intercept = Intercept - dot_product(means_X, b);
#> }

# define a multi-normal prior with known covariance matrix
bprior <- prior(multi_normal(M, V), class = "b")
stanvars <- stanvar(rep(0, 2), "M", scode = "  vector[K] M;") +
  stanvar(diag(2), "V", scode = "  matrix[K, K] V;")
make_stancode(count ~ Trt + zBase, epilepsy,
              prior = bprior, stanvars = stanvars)

#> // generated with brms 2.10.3
#> functions {
#> }
#> data {
#>   int<lower=1> N;  // number of observations
#>   vector[N] Y;  // response variable
#>   int<lower=1> K;  // number of population-level effects
#>   matrix[N, K] X;  // population-level design matrix
#>   int prior_only;  // should the likelihood be ignored?
#>   vector[K] M;
#>   matrix[K, K] V;
#> }
#> transformed data {
#>   int Kc = K - 1;
#>   matrix[N, Kc] Xc;  // centered version of X without an intercept
#>   vector[Kc] means_X;  // column means of X before centering
#>   for (i in 2:K) {
#>     means_X[i - 1] = mean(X[, i]);
#>     Xc[, i - 1] = X[, i] - means_X[i - 1];
#>   }
#> }
#> parameters {
#>   vector[Kc] b;  // population-level effects
#>   // temporary intercept for centered predictors
#>   real Intercept;
#>   real<lower=0> sigma;  // residual SD
#> }
#> transformed parameters {
#> }
#> model {
#>   // priors including all constants
#>   target += multi_normal_lpdf(b | M, V);
#>   target += student_t_lpdf(Intercept | 3, 4, 10);
#>   target += student_t_lpdf(sigma | 3, 0, 10)
#>     - 1 * student_t_lccdf(0 | 3, 0, 10);
#>   // likelihood including all constants
#>   if (!prior_only) {
#>     target += normal_id_glm_lpdf(Y | Xc, Intercept, b, sigma);
#>   }
#> }
#> generated quantities {
#>   // actual population-level intercept
#>   real b_Intercept = Intercept - dot_product(means_X, b);
#> }

# define a hierachical prior on the regression coefficients
bprior <- set_prior("normal(0, tau)", class = "b") +
  set_prior("target += normal_lpdf(tau | 0, 10)", check = FALSE)
stanvars <- stanvar(scode = "real<lower=0> tau;",
                    block = "parameters")
make_stancode(count ~ Trt + zBase, epilepsy,
              prior = bprior, stanvars = stanvars)

#> // generated with brms 2.10.3
#> functions {
#> }
#> data {
#>   int<lower=1> N;  // number of observations
#>   vector[N] Y;  // response variable
#>   int<lower=1> K;  // number of population-level effects
#>   matrix[N, K] X;  // population-level design matrix
#>   int prior_only;  // should the likelihood be ignored?
#> }
#> transformed data {
#>   int Kc = K - 1;
#>   matrix[N, Kc] Xc;  // centered version of X without an intercept
#>   vector[Kc] means_X;  // column means of X before centering
#>   for (i in 2:K) {
#>     means_X[i - 1] = mean(X[, i]);
#>     Xc[, i - 1] = X[, i] - means_X[i - 1];
#>   }
#> }
#> parameters {
#>   vector[Kc] b;  // population-level effects
#>   // temporary intercept for centered predictors
#>   real Intercept;
#>   real<lower=0> sigma;  // residual SD
#> real<lower=0> tau;
#> }
#> transformed parameters {
#> }
#> model {
#>   // priors including all constants
#>   target += normal_lpdf(b | 0, tau);
#>   target += student_t_lpdf(Intercept | 3, 4, 10);
#>   target += student_t_lpdf(sigma | 3, 0, 10)
#>     - 1 * student_t_lccdf(0 | 3, 0, 10);
#>   target += normal_lpdf(tau | 0, 10);
#>   // likelihood including all constants
#>   if (!prior_only) {
#>     target += normal_id_glm_lpdf(Y | Xc, Intercept, b, sigma);
#>   }
#> }
#> generated quantities {
#>   // actual population-level intercept
#>   real b_Intercept = Intercept - dot_product(means_X, b);
#> }