Black Theme for ggplot2 Graphics

A black theme for ggplot graphics inspired by a blog post of Jon Lefcheck (https://jonlefcheck.net/2013/03/11/black-theme-for-ggplot2-2/).

theme_black(base_size = 12, base_family = "")

Arguments

base_size
base font size

base_family
base font family

Value

A theme object used in ggplot2 graphics.

Details

When using theme_black in plots powered by the bayesplot package such as pp_check or stanplot, I recommend using the "viridisC" color scheme (see examples).

Examples

# NOT RUN {
# change default ggplot theme
ggplot2::theme_set(theme_black())

# change default bayesplot color scheme
bayesplot::color_scheme_set("viridisC")

# fit a simple model
fit <- brm(count ~ zAge + zBase * Trt + (1|patient),
           data = epilepsy, family = poisson(), chains = 2)
summary(fit)

# create various plots
plot(marginal_effects(fit), ask = FALSE)
pp_check(fit)
stanplot(fit, type = "hex", pars = c("b_Intercept", "b_Trt1"))
# }