vcov.brmsfit {brms}	R Documentation

Covariance and Correlation Matrix of Population-Level Effects

Description

Get a point estimate of the covariance or correlation matrix of population-level parameters

Usage

## S3 method for class 'brmsfit'
vcov(object, correlation = FALSE, ...)

Arguments

object
An object of class brmsfit

correlation
logical; if FALSE (the default), compute the covariance matrix, if TRUE, compute the correlation matrix

...
Currently ignored

Details

Estimates are obtained by calculating the maximum likelihood covariances (correlations) of the posterior samples.

Value

covariance or correlation matrix of population-level parameters

Examples

## Not run: 
fit <- brm(count ~ log_Age_c + log_Base4_c * Trt_c + (1+Trt_c|visit), 
           data = epilepsy, family = gaussian(), chains = 2)
vcov(fit)

## End(Not run)

[Package brms version 2.3.0 Index]